requirements.txt
fpdf==1.7.2
python-dotenv==1.0.1
Requests==2.32.3
